shiiish, now u can search, no need to do two actions like "add, delete", now it handled by one action. isFriend now pass with user.

+++ api/v1/user/views.py
@@ -1,5 +1,5 @@
 from rest_framework.decorators import action
-from rest_framework.permissions import AllowAny
+from rest_framework.permissions import AllowAny, IsAuthenticated
 
 from apps.post.models import Post
 from apps.post.serializers import PostSerializer
@@ -7,17 +7,19 @@ from ..abstract.views import AbstractViewSet
 from apps.user.serializers import UserSerializer
 from apps.user.models import User
 from rest_framework.response import Response
-
+from rest_framework import filters
 
 class UserViewSet(AbstractViewSet):
     http_method_names = ('patch', 'get', 'post')
-    permission_classes = (AllowAny,)
+    permission_classes = (IsAuthenticated,)
+    filter_backends = [filters.SearchFilter]
+    search_fields = ['username']
     serializer_class = UserSerializer
 
     def get_queryset(self):
         if not self.request.user.is_anonymous and self.request.user.is_superuser:
             return User.objects.all()
-        return User.objects.exclude(is_superuser=True)
+        return User.objects.exclude(is_superuser=True).exclude(public_id=self.request.user.public_id)
 
     def get_object(self):
         obj = User.objects.get_object_by_public_id(self.kwargs['pk'])
@@ -27,20 +29,19 @@ class UserViewSet(AbstractViewSet):
     @action(methods=['get'], detail=True)
     def newsfeed(self, request, *args, **kwargs):
         user = self.get_object()
-        posts = Post.objects.filter(author__in=user.friends.all()).order_by('-created')
+        posts = Post.objects.filter(author__in=user.friends.all()).order_by('-created') | user.posts.all().order_by('-created')
         page = self.paginate_queryset(posts)
         if page is not None:
-            serializer = PostSerializer(page, many=True)
+            serializer = PostSerializer(page, context={'request': request}, many=True)
             return self.get_paginated_response(serializer.data)
-        serializer = PostSerializer(posts, many=True)
+        serializer = PostSerializer(posts, context={'request': request}, many=True)
         return Response(serializer.data)
-
 
     @action(methods=['get'], detail=True)
     def friends(self, request, *args, **kwargs):
         user = self.get_object()
         friends = user.friends.all()
-        friends_serializer = UserSerializer(friends, many=True)
+        friends_serializer = UserSerializer(friends, context={'request': request}, many=True)
         return Response(friends_serializer.data)
 
     @action(methods=['get'], detail=True)
@@ -48,9 +49,9 @@ class UserViewSet(AbstractViewSet):
         user = self.get_object()
         serializer = self.get_serializer(user)
         friends = user.friends.all()
-        friends_serializer = UserSerializer(friends, many=True)
+        friends_serializer = UserSerializer(friends, context={'request': request}, many=True)
         posts = user.posts.all()
-        posts_serializer = PostSerializer(posts, many=True)
+        posts_serializer = PostSerializer(posts, context={'request': request},  many=True)
         output = {
             "user": serializer.data,
             "friends": friends_serializer.data,
@@ -62,18 +63,11 @@ class UserViewSet(AbstractViewSet):
     def add_friend(self, request, *args, **kwargs):
         friend = self.get_object()
         user = self.request.user
-
-        user.add_friend(friend)
-        serializer = self.serializer_class(friend)
-        return Response(serializer.data)
-
-    @action(methods=['post'], detail=True)
-    def delete_friend(self, request, *args, **kwargs):
-        friend = self.get_object()
-        user = self.request.user
-
-        user.delete_friend(friend)
-        serializer = self.serializer_class(friend)
+        if user.friends.filter(public_id = friend.public_id).exists():
+            user.delete_friend(friend)
+        else:
+            user.add_friend(friend)
+        serializer = self.serializer_class(friend, context={"request": request})
         return Response(serializer.data)
 
     @action(methods=['get'], detail=True)
